Moves "extended-table" plugin into "plugins/table" package.
